* release: (2515 commits)
  develop: bump up salt/cloud/deploy/bootstrap-salt.sh to the latest release
  develop: add get_pubip util to grains and introspect module, duplicate ini_manager to iniconf
  develop: update .travis.yml to add notifications to slack devnotice channel
  develop: add skip_verify option to file.managed states module with a simple presetting trick
  develop: replace all saltstack bootstrap-salt.sh with github mubiic/saltstrap/master/bootstrap-salt.sh
  Added known issue #30820 to 2015.8.4 release notes
  Updated man pages, version to 2015.8.4, added CVE 2016-1866 to release notes
  Fix some autodoc import warnings from building docs
  Issue #28706: Fix state user.present behavior.
  Fix for win_network changing dhcp to static DNS
  Fix password is treated as int instead of string in postgres module
  Add new dimension data to doc list
  Make `salt-ssh` executions less of an exception black hole
  Teach python 2.6 how to copy the object.
  Lint fixes
  Lint fixes
  Added ability to install and activate a windows license
  Added the ability to add and remove certificates from Windows minions
  Updated based on PR feedback
  extra space before disabling import warning
  ...

+++ salt/utils/parsers.py
@@ -15,6 +15,7 @@
 from __future__ import absolute_import, print_function
 import os
 import sys
+import types
 import signal
 import getpass
 import logging
@@ -130,7 +131,7 @@ class OptionParser(optparse.OptionParser, object):
     usage = '%prog'
 
     epilog = ('You can find additional help about %prog issuing "man %prog" '
-              'or on http://docs.saltstack.org')
+              'or on http://docs.saltstack.com')
     description = None
 
     # Private attributes
@@ -783,6 +784,7 @@ class LogLevelMixIn(six.with_metaclass(MixInMeta, object)):
     def _setup_mp_logging_listener(self, *args):  # pylint: disable=unused-argument
         if self._setup_mp_logging_listener_:
             log.setup_multiprocessing_logging_listener(
+                self.config,
                 self._get_mp_logging_listener_queue()
             )
 
@@ -867,8 +869,11 @@ class DaemonMixIn(six.with_metaclass(MixInMeta, object)):
         )
 
     def _mixin_before_exit(self):
-        if self.check_pidfile():
-            os.unlink(self.config['pidfile'])
+        if hasattr(self, 'config'):
+            # We've loaded and merged options into the configuration, it's safe
+            # to query about the pidfile
+            if self.check_pidfile():
+                os.unlink(self.config['pidfile'])
 
     def set_pidfile(self):
         from salt.utils.process import set_pidfile
@@ -893,7 +898,7 @@ class DaemonMixIn(six.with_metaclass(MixInMeta, object)):
             if self._setup_mp_logging_listener_ is True:
                 # Stop the logging queue listener for the current process
                 # We'll restart it once forked
-                log.shutdown_multiprocessing_logging_listener()
+                log.shutdown_multiprocessing_logging_listener(daemonizing=True)
 
             # Late import so logging works correctly
             salt.utils.daemonize()
@@ -929,10 +934,11 @@ class DaemonMixIn(six.with_metaclass(MixInMeta, object)):
         self._install_signal_handlers()
 
     def _handle_signals(self, signum, sigframe):  # pylint: disable=unused-argument
+        msg = self.__class__.__name__
         if signum == signal.SIGINT:
-            msg = 'Received a SIGINT.'
+            msg += ' received a SIGINT.'
         elif signum == signal.SIGTERM:
-            msg = 'Received a SIGTERM.'
+            msg += ' received a SIGTERM.'
         logging.getLogger(__name__).warning('{0} Exiting.'.format(msg))
         self.shutdown(exitmsg='{0} Exited.'.format(msg))
 
@@ -950,31 +956,46 @@ class PidfileMixin(six.with_metaclass(MixInMeta, object)):
             'DaemonMixIn which contains the same behavior. PidfileMixin '
             'will be supported until Salt {version}.'
         )
-        self.add_option(
-            '--pid-file', dest='pidfile',
-            default=os.path.join(
-                syspaths.PIDFILE_DIR, '{0}.pid'.format(self.get_prog_name())
-            ),
-            help=('Specify the location of the pidfile. Default: %default')
-        )
+        try:
+            self.add_option(
+                '--pid-file', dest='pidfile',
+                default=os.path.join(
+                    syspaths.PIDFILE_DIR, '{0}.pid'.format(self.get_prog_name())
+                ),
+                help=('Specify the location of the pidfile. Default: %default')
+            )
 
-    def set_pidfile(self):
-        from salt.utils.process import set_pidfile
-        set_pidfile(self.config['pidfile'], self.config['user'])
+            # Since there was no colision with DaemonMixin, let's add the
+            # pidfile mixin methods. This is used using types.MethodType
+            # because if we had defined these at the class level, they would
+            # have overridden the exact same methods from the DaemonMixin.
 
-    def check_pidfile(self):
-        '''
-        Report whether a pidfile exists
-        '''
-        from salt.utils.process import check_pidfile
-        return check_pidfile(self.config['pidfile'])
+            def set_pidfile(self):
+                from salt.utils.process import set_pidfile
+                set_pidfile(self.config['pidfile'], self.config['user'])
 
-    def get_pidfile(self):
-        '''
-        Return a pid contained in a pidfile
-        '''
-        from salt.utils.process import get_pidfile
-        return get_pidfile(self.config['pidfile'])
+            self.set_pidfile = types.MethodType(set_pidfile, self)
+
+            def check_pidfile(self):
+                '''
+                Report whether a pidfile exists
+                '''
+                from salt.utils.process import check_pidfile
+                return check_pidfile(self.config['pidfile'])
+
+            self.check_pidfile = types.MethodType(check_pidfile, self)
+
+            def get_pidfile(self):
+                '''
+                Return a pid contained in a pidfile
+                '''
+                from salt.utils.process import get_pidfile
+                return get_pidfile(self.config['pidfile'])
+
+            self.get_pidfile = types.MethodType(get_pidfile, self)
+        except optparse.OptionConflictError:
+            # The option was already added by the DaemonMixin
+            pass
 
 
 class TargetOptionsMixIn(six.with_metaclass(MixInMeta, object)):
@@ -1339,7 +1360,7 @@ class ExecutionOptionsMixIn(six.with_metaclass(MixInMeta, object)):
             '-m', '--map',
             default=None,
             help='Specify a cloud map file to use for deployment. This option '
-                 'may be used alone, or in conjunction with -Q, -F, -S or -d.'
+                 'may be used alone, or in conjunction with -Q, -F, -S or -d. '
                  'The map can also be filtered by a list of VM names.'
         )
         group.add_option(
@@ -1809,6 +1830,14 @@ class SaltCMDOptionParser(six.with_metaclass(OptionParserMeta,
             help=('Execute the salt job in batch mode, pass either the number '
                   'of minions to batch at a time, or the percentage of '
                   'minions to have running')
+        )
+        self.add_option(
+            '--batch-wait',
+            default=0,
+            dest='batch_wait',
+            type=float,
+            help=('Wait the specified time in seconds after each job is done '
+                  'before freeing the slot in the batch for the next one')
         )
         self.add_option(
             '-a', '--auth', '--eauth', '--external-auth',
@@ -2684,7 +2713,7 @@ class SaltSSHOptionParser(six.with_metaclass(OptionParserMeta,
             '-w', '--wipe',
             default=False,
             action='store_true',
-            dest='wipe_ssh',
+            dest='ssh_wipe',
             help='Remove the deployment of the salt files when done executing.',
         )
         self.add_option(
@@ -2889,7 +2918,8 @@ class SPMParser(six.with_metaclass(OptionParserMeta,
                                    OptionParser,
                                    ConfigDirMixIn,
                                    LogLevelMixIn,
-                                   MergeConfigMixIn)):
+                                   MergeConfigMixIn,
+                                   SaltfileMixIn)):
     '''
     The cli parser object used to fire up the salt spm system.
     '''
@@ -2915,6 +2945,12 @@ class SPMParser(six.with_metaclass(OptionParserMeta,
             action='store_true',
             help='Default yes in answer to all confirmation questions.'
         )
+        self.add_option(
+            '-v', '--verbose',
+            default=False,
+            action='store_true',
+            help='Display more detailed information.'
+        )
 
     def _mixin_after_parsed(self):
         # spm needs arguments
@@ -2925,3 +2961,21 @@ class SPMParser(six.with_metaclass(OptionParserMeta,
 
     def setup_config(self):
         return salt.config.spm_config(self.get_config_file_path())
+
+
+class SaltAPIParser(six.with_metaclass(OptionParserMeta,
+                                       OptionParser,
+                                       ConfigDirMixIn,
+                                       LogLevelMixIn,
+                                       DaemonMixIn,
+                                       MergeConfigMixIn)):
+    '''
+    The Salt API cli parser object used to fire up the salt api system.
+    '''
+    # ConfigDirMixIn config filename attribute
+    _config_filename_ = 'master'
+    # LogLevelMixIn attributes
+    _default_logging_logfile_ = os.path.join(syspaths.LOGS_DIR, 'api')
+
+    def setup_config(self):
+        return salt.config.api_config(self.get_config_file_path())  # pylint: disable=no-member
